Rename `kicker` to `tagline`.
